test(proving-ground): promote PTY harness into shared support module
