+++ crates/wcore-cli/tests/smoke_p0.rs
@@ -84,6 +84,7 @@ mod support;
 use std::path::Path;
 
 use support::mock_llm::{MockLlm, RecordedRequest, received_requests};
+use support::pty::{harden_child_env, write_config};
 use tempfile::TempDir;
 use wiremock::MockServer;
 
@@ -94,23 +95,6 @@ use wiremock::MockServer;
 /// Path to the debug binary under test (Cargo wires this env var).
 fn binary() -> &'static str {
     env!("CARGO_BIN_EXE_wayland-core")
-}
-
-/// Seed `<home>/config.toml` for a provider/model, optionally routing the
-/// provider `base_url` at a local mock. `model: None` writes NO model line —
-/// the exact catalog-provider shape the D002 GAP check needs.
-fn write_config(home: &Path, provider: &str, model: Option<&str>, base_url: Option<&str>) {
-    let mut toml = format!("[default]\nprovider = \"{provider}\"\n");
-    if let Some(m) = model {
-        toml.push_str(&format!("model = \"{m}\"\n"));
-    }
-    toml.push_str(&format!(
-        "\n[providers.{provider}]\napi_key = \"sk-ant-harness-not-real-key-0000000000\"\n"
-    ));
-    if let Some(url) = base_url {
-        toml.push_str(&format!("base_url = \"{url}\"\n"));
-    }
-    std::fs::write(home.join("config.toml"), toml).expect("write config.toml");
 }
 
 /// Start a [`MockServer`] on a held tokio runtime, returning both so the caller
@@ -126,50 +110,6 @@ fn start_mock(mock: MockLlm) -> (tokio::runtime::Runtime, MockServer) {
 /// was started on. The rebind-class checks assert on these.
 fn recorded(rt: &tokio::runtime::Runtime, server: &MockServer) -> Vec<RecordedRequest> {
     rt.block_on(received_requests(server))
-}
-
-/// The full provider-credential env-var set every spawned child must NOT
-/// inherit, so a run can neither read the developer's real keys nor have
-/// onboarding auto-detect a stray dev credential. ONE source of truth used by
-/// `run_headless`, the PTY spawn, and the `--json-stream` child — keeps the
-/// strip set honest and uniform (M6). `AWS_*` / `VERTEX*` are stripped by name
-/// (the concrete vars Bedrock/Vertex auth reads), not by glob.
-const STRIPPED_PROVIDER_ENV: &[&str] = &[
-    "API_KEY",
-    "ANTHROPIC_API_KEY",
-    "OPENAI_API_KEY",
-    "GEMINI_API_KEY",
-    "GOOGLE_API_KEY",
-    "OPENROUTER_API_KEY",
-    "DEEPSEEK_API_KEY",
-    "GROQ_API_KEY",
-    // AWS (Bedrock) — concrete vars the provider auth chain reads.
-    "AWS_ACCESS_KEY_ID",
-    "AWS_SECRET_ACCESS_KEY",
-    "AWS_SESSION_TOKEN",
-    "AWS_PROFILE",
-    "AWS_REGION",
-    "AWS_DEFAULT_REGION",
-    // Google Vertex.
-    "VERTEX_PROJECT",
-    "VERTEX_LOCATION",
-    "GOOGLE_APPLICATION_CREDENTIALS",
-];
-
-/// Apply the hermetic child env uniformly: point `WAYLAND_HOME` + `HOME` at the
-/// throwaway tempdir, set a deterministic `TERM`, and strip every credential in
-/// [`STRIPPED_PROVIDER_ENV`]. The single place that defines "hermetic child
-/// env" so the headless / PTY / json-stream spawns can never drift apart (M6).
-fn harden_child_env(cmd: &mut std::process::Command, home: &Path) {
-    cmd.env("WAYLAND_HOME", home)
-        .env("HOME", home)
-        // Headless / json-stream children get a deterministic non-TTY term. The
-        // PTY spawn (which needs a real terminal type) sets its own TERM and
-        // does NOT route through this helper.
-        .env("TERM", "dumb");
-    for key in STRIPPED_PROVIDER_ENV {
-        cmd.env_remove(key);
-    }
 }
 
 /// Spawn the binary headless (no TUI) against a hermetic home. The prompt is a
@@ -400,140 +340,10 @@ fn smoke_17_force_posture_auto_approves_mutating_tool_in_engine() {
 
 #[cfg(unix)]
 mod pty {
-    use super::*;
-    use std::io::{Read, Write};
     use std::time::{Duration, Instant};
 
-    use portable_pty::{CommandBuilder, MasterPty, PtySize, native_pty_system};
-
-    /// A minimal PTY harness — the proven shape from `harness_tui_flow.rs`,
-    /// re-derived here because integration test files compile as separate
-    /// binaries and cannot share a non-`support` module.
-    pub struct Pty {
-        writer: Box<dyn Write + Send>,
-        parser: std::sync::Arc<std::sync::Mutex<vt100::Parser>>,
-        _master: Box<dyn MasterPty + Send>,
-        child: Box<dyn portable_pty::Child + Send + Sync>,
-        _reader: std::thread::JoinHandle<()>,
-    }
-
-    impl Pty {
-        pub fn spawn(home: &Path) -> Self {
-            let pty = native_pty_system()
-                .openpty(PtySize {
-                    rows: 40,
-                    cols: 120,
-                    pixel_width: 0,
-                    pixel_height: 0,
-                })
-                .expect("open PTY");
-
-            let mut cmd = CommandBuilder::new(binary());
-            cmd.env("HOME", home);
-            cmd.env("WAYLAND_HOME", home);
-            // The TUI needs a real terminal type (not "dumb") to render; the
-            // hermetic key-strip set is shared with the headless/json-stream
-            // spawns via STRIPPED_PROVIDER_ENV (M6).
-            cmd.env("TERM", "xterm-256color");
-            for key in STRIPPED_PROVIDER_ENV {
-                cmd.env_remove(key);
-            }
-            cmd.cwd(home);
-            let child = pty.slave.spawn_command(cmd).expect("spawn wayland-core");
-
-            let mut reader = pty.master.try_clone_reader().expect("clone PTY reader");
-            let parser = std::sync::Arc::new(std::sync::Mutex::new(vt100::Parser::new(40, 120, 0)));
-            let parser_for_thread = std::sync::Arc::clone(&parser);
-            let reader_handle = std::thread::spawn(move || {
-                let mut buf = [0u8; 8192];
-                loop {
-                    match reader.read(&mut buf) {
-                        Ok(0) => break,
-                        Ok(n) => {
-                            if let Ok(mut p) = parser_for_thread.lock() {
-                                p.process(&buf[..n]);
-                            }
-                        }
-                        Err(_) => break,
-                    }
-                }
-            });
-
-            let writer = pty.master.take_writer().expect("take PTY writer");
-            Self {
-                writer,
-                parser,
-                _master: pty.master,
-                child,
-                _reader: reader_handle,
-            }
-        }
-
-        pub fn screen_text(&self) -> String {
-            let parser = self.parser.lock().expect("parser lock");
-            parser.screen().contents()
-        }
-
-        pub fn wait_for<F: Fn(&str) -> bool>(&self, predicate: F, timeout: Duration, what: &str) {
-            let deadline = Instant::now() + timeout;
-            let mut last = String::new();
-            while Instant::now() < deadline {
-                last = self.screen_text();
-                if predicate(&last) {
-                    return;
-                }
-                std::thread::sleep(Duration::from_millis(30));
-            }
-            panic!(
-                "timed out after {:?} waiting for {what}.\n--- last screen ---\n{}\n--- end ---",
-                timeout, last
-            );
-        }
-
-        pub fn send(&mut self, bytes: &[u8]) {
-            self.writer.write_all(bytes).expect("write to PTY");
-            self.writer.flush().ok();
-        }
-
-        pub fn wait_for_exit(&mut self, timeout: Duration) -> Option<portable_pty::ExitStatus> {
-            let deadline = Instant::now() + timeout;
-            while Instant::now() < deadline {
-                match self.child.try_wait() {
-                    Ok(Some(status)) => return Some(status),
-                    Ok(None) => std::thread::sleep(Duration::from_millis(50)),
-                    Err(_) => return None,
-                }
-            }
-            None
-        }
-
-        /// Clean shutdown via the proven palette quit path.
-        pub fn quit(&mut self) {
-            self.send(b"/");
-            std::thread::sleep(Duration::from_millis(300));
-            self.send(b"exit\r");
-            let _ = self.wait_for_exit(Duration::from_secs(8));
-        }
-    }
-
-    impl Drop for Pty {
-        fn drop(&mut self) {
-            if let Ok(None) = self.child.try_wait() {
-                let _ = self.child.kill();
-            }
-        }
-    }
-
-    /// Boot the TUI to the Workspace surface (chrome wordmark + tab painted).
-    pub fn boot(home: &Path) -> Pty {
-        let h = Pty::spawn(home);
-        h.wait_for(
-            |s| s.contains("WAYLAND") && s.contains("Workspace"),
-            Duration::from_secs(60),
-            "TUI to render the chrome wordmark and Workspace tab",
-        );
-        h
-    }
+    use super::{MockLlm, TempDir, start_mock, write_config};
+    use crate::support::pty::{Pty, boot};
 
     // -------------------------------------------------------------------
     // SMOKE #24 — `/quit` clean exit (the always-on sanity anchor). Cheap,
