+++ crates/wcore-cli/tests/support/mod.rs
@@ -3,3 +3,4 @@
 //! Included from integration test files via `#[path = "support/mod.rs"] mod support;`.
 
 pub mod mock_llm;
+pub mod pty;
